requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
vcfpy
google-generativeai>=0.8.0